requirements.txt
torch>=2.1.0
torchvision>=0.16.0
monai>=1.3.0
monai-generative>=0.2.3
nibabel>=5.0.0
numpy>=1.24.0
scipy>=1.11.0
SimpleITK>=2.3.0
scikit-image>=0.21.0
matplotlib>=3.7.0
pyyaml>=6.0
tqdm>=4.65.0
einops>=0.7.0
# optional for radiomics quality check
pyradiomics>=3.1.0
